fuels-code-gen extracted, code compiling tests passing
