+++ packages/fuels-code-gen/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "fuels-code-gen"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+description = "Used for code generation in the Fuel Rust SDK"
+
+[dependencies]
+Inflector = "0.11.4"
+fuel-abi-types = "0.2.0"
+itertools = "0.10.5"
+proc-macro2 = "1.0.50"
+quote = "1.0.23"
+serde_json = "1.0.91"
+lazy_static = "1.4.0"
+regex = "1.7.1"
+syn = { version = "1.0.107" }
