+++ packages/fuels-code-gen/src/lib.rs
@@ -1,0 +1,5 @@
+mod error;
+mod program_bindings;
+pub mod utils;
+
+pub use program_bindings::*;
