+++ packages/fuels-code-gen/src/program_bindings.rs
@@ -1,0 +1,9 @@
+mod abi_types;
+mod abigen;
+mod custom_types;
+mod generated_code;
+mod resolved_type;
+mod source;
+mod utils;
+
+pub use abigen::{Abigen, AbigenTarget, ProgramType};
