+++ packages/fuels-code-gen/src/program_bindings/abi_types.rs
