+++ packages/fuels-code-gen/src/program_bindings/abigen/abigen_target.rs
@@ -1,8 +1,8 @@
 use std::convert::TryFrom;
 
 use crate::{
-    abigen::code_gen::{abi_types::FullProgramABI, source::Source},
     error::{Error, Result},
+    program_bindings::{abi_types::FullProgramABI, source::Source},
 };
 
 #[derive(Debug, Clone)]
