+++ packages/fuels-code-gen/src/program_bindings/abigen/bindings/script.rs
@@ -3,17 +3,17 @@ use std::collections::HashSet;
 use proc_macro2::{Ident, TokenStream};
 use quote::quote;
 
+use crate::utils::type_path::TypePath;
 use crate::{
-    abigen::code_gen::{
+    error::Result,
+    program_bindings::{
         abi_types::{FullProgramABI, FullTypeDeclaration},
         abigen::{
             bindings::{function_generator::FunctionGenerator, utils::extract_main_fn},
             logs::logs_lookup_instantiation_code,
         },
         generated_code::GeneratedCode,
-        type_path::TypePath,
     },
-    error::Result,
 };
 
 pub(crate) fn script_bindings(
