+++ packages/fuels-code-gen/src/program_bindings/code_gen.rs
