+++ packages/fuels-code-gen/src/program_bindings/custom_types/structs.rs
@@ -4,15 +4,15 @@ use fuel_abi_types::utils::extract_custom_type_name;
 use proc_macro2::{Ident, TokenStream};
 use quote::quote;
 
+use crate::utils::type_path::TypePath;
 use crate::{
-    abigen::code_gen::{
+    error::{error, Result},
+    program_bindings::{
         abi_types::FullTypeDeclaration,
         custom_types::utils::{extract_components, extract_generic_parameters},
         generated_code::GeneratedCode,
-        type_path::TypePath,
         utils::Component,
     },
-    error::{error, Result},
     utils::ident,
 };
