+++ packages/fuels-code-gen/src/program_bindings/source.rs
