+++ packages/fuels-code-gen/src/utils.rs
@@ -1,0 +1,4 @@
+mod ident;
+pub mod type_path;
+
+pub use ident::*;
