+++ packages/fuels-code-gen/src/utils/ident.rs
@@ -1,0 +1,10 @@
+use proc_macro2::{Ident, Span};
+
+/// Expands a identifier string into an token.
+pub fn ident(name: &str) -> Ident {
+    Ident::new(name, Span::call_site())
+}
+
+pub fn safe_ident(name: &str) -> Ident {
+    syn::parse_str::<Ident>(name).unwrap_or_else(|_| ident(&format!("{name}_")))
+}
