+++ packages/fuels-macros/src/abigen.rs
@@ -1,5 +1,3 @@
-mod code_gen;
-mod parsing;
-
-pub(crate) use code_gen::{Abigen, AbigenTarget, ProgramType, TypePath};
 pub(crate) use parsing::MacroAbigenTargets;
+
+mod parsing;
