--- packages/fuels-macros/src/abigen/code_gen.rs
@@ -1,13 +1,0 @@
-//! This module implements everything related to code generation/expansion
-//! from a FuelVM ABI.
-mod abi_types;
-mod abigen;
-mod custom_types;
-mod generated_code;
-mod resolved_type;
-mod source;
-mod type_path;
-mod utils;
-
-pub(crate) use abigen::{Abigen, AbigenTarget, ProgramType};
-pub(crate) use type_path::TypePath;
